Wo 155 Usecases appear below the device selection
This is a combination of 28 commits.
Moved smart clipping images to their separate folder

made base.js -> base.js.coffee

Extracted devices view in a template

Added Underscore to simplify parts of code

List of available usecases for the selected devices is now dynamic, depending on selected devices.

Added showcaseManager to take care o showing and hiding of the animations depending on the selected devices

renamed folders

Renamed folders, added latest version of fullpage and slimscroll.
Fullpage control is reinitialized when showcases show up.

Added info on event objects

extracted functionality in DevicesManager that handles the showing of the devices and returning the selected devices. and Showcase manager that shows the usecases for the selected devices.

Replaced and organized animations

Updated animations

Animations appear under the device selection and the naming doesn't matter anymore

Animations are now starting up when they are 50% in the view.

Animations are played on scroll only once

Replay button is hidden at the start of the animation and only shows up when the animation is finished

Refactored parts of the showcase_presenter to encapsulate animation functionality into a javascript object

Automatic scrolling to the first animation

Menu is now changing color depending on the background behind it

Title bar shows up below the menu

fixed the fullpage sections

See more button is working

Added jQuery.nearest file

Fixed merge

Scroll to usecases looks like the scroll to another section

Updated animations

Added replay and more info buttons.
Centered the animations
Fixed problem when fullpage is not used (on mobile devices)

Added the more info modal in the usecase_template